test-requirements.txt
testrepository
hacking
coverage
requests-mock
testtools
stestr
mock
